chore(bookings): updated action and types
